[tasks] Rename pages as panels
We want to reuse these panels later to form pages
